feat: org re-export sdk-wallet
