+++ lerna.json
@@ -1,18 +1,10 @@
 {
-  "packages": [
-    "packages/**/*"
-  ],
+  "packages": ["packages/**/*"],
   "version": "0.13.68-fix.23",
   "npmClient": "yarn",
-  "useWorkspaces": true,
   "command": {
     "publish": {
-      "allowBranch": [
-        "master",
-        "fix/*",
-        "beta/*",
-        "release/*"
-      ],
+      "allowBranch": ["master", "fix/*", "beta/*", "release/*"],
       "conventionalCommits": true,
       "registry": "https://registry.npmjs.org/",
       "message": "chore: bump versions after release",
