+++ package.json
@@ -5,9 +5,9 @@
     "packages/**/*"
   ],
   "scripts": {
-    "bootstrap": "lerna bootstrap && lerna link",
+    "bootstrap": "corepack yarn install",
     "clean": "lerna run clean",
-    "build": "run-s bootstrap clean build:lerna",
+    "build": "run-s clean build:lerna",
     "build:lerna": "lerna run build --stream --ignore example --ignore @rarible/sdk-examples",
     "build-ethereum-test-common": "yarn workspace @rarible/ethereum-sdk-test-common run build",
     "build-ethereum-biconomy-middleware": "yarn workspace @rarible/biconomy-middleware run build",
@@ -106,7 +106,7 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-junit": "^15.0.0",
-    "lerna": "^6.6.2",
+    "lerna": "^7.4.2",
     "lint-staged": "^10.4.0",
     "magic-sdk": "~18.4.1",
     "npm-check-updates": "^11.5.13",
@@ -117,8 +117,13 @@
     "rxjs": "^6.6.7",
     "sort-package-json": "^1.52.0",
     "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
     "tslib": "^2.3.1",
     "typescript": "^5.4.5",
     "web3": "1.8.2"
+  },
+  "packageManager": "yarn@1.22.22",
+  "engines": {
+    "node": ">=18.0.0"
   }
 }
