+++ packages/sdk/src/index.ts
@@ -1,10 +1,9 @@
 import type { ContractAddress, Maybe } from "@rarible/types"
 import type { CollectionId } from "@rarible/api-client"
 import { Blockchain } from "@rarible/api-client"
-import type { BlockchainWallet, WalletByBlockchain } from "@rarible/sdk-wallet"
+import type { BlockchainWallet, WalletByBlockchain, RaribleSdkProvider } from "@rarible/sdk-wallet/build"
 import { WalletType } from "@rarible/sdk-wallet"
 import { getRandomId } from "@rarible/utils"
-import type { RaribleSdkProvider } from "@rarible/sdk-wallet/build/get-wallet"
 import { getRaribleWallet } from "@rarible/sdk-wallet/build/get-wallet"
 import type { AbstractLogger } from "@rarible/logger/build/domain"
 import type { SupportedBlockchain } from "@rarible/sdk-common"
