--- packages/wallet/jest.config.js
@@ -1,12 +1,0 @@
-module.exports = {
-  roots: ["<rootDir>/src"],
-  transform: {
-    "^.+\\.ts?$": [
-      "ts-jest",
-      {
-        tsconfig: "tsconfig-build.json",
-      },
-    ],
-  },
-  transformIgnorePatterns: ["<rootDir>/build/", "<rootDir>/node_modules/"],
-}
