+++ packages/wallet/jest.config.ts
@@ -1,0 +1,9 @@
+import type { JestConfigWithTsJest } from "ts-jest"
+
+const jestConfig: JestConfigWithTsJest = {
+  roots: ["<rootDir>/src"],
+  moduleNameMapper: { "^(\\.{1,2}/.*)\\.js$": "$1" },
+  transform: { "^.+\\.ts?$": "ts-jest" },
+}
+
+export default jestConfig
